requirements.txt
discord.py
yt-dlp
PyNaCl
python-dotenv
lyricsgenius
openai
aiohttp
davey
